Add, modify and delete company

+++ src/main/java/com/app/advert/portal/service/impl/UserServiceImpl.java
@@ -1,16 +1,19 @@
 package com.app.advert.portal.service.impl;
 
 import com.app.advert.portal.dto.UserDto;
+import com.app.advert.portal.enums.UserRole;
 import com.app.advert.portal.mapper.UserMapper;
 import com.app.advert.portal.model.Role;
 import com.app.advert.portal.model.User;
+import com.app.advert.portal.security.SecurityUtils;
 import com.app.advert.portal.service.UserService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
 import java.util.List;
 
 
@@ -24,13 +27,12 @@ public class UserServiceImpl implements UserService {
     private final PasswordEncoder passwordEncoder;
 
     @Override
-    public List<User> getAll() {
-        return userMapper.getAll();
-    }
-
-    @Override
-    public User getById(Long id) {
-        return userMapper.getById(id);
+    public ResponseEntity getById(Long id) {
+        User user = userMapper.getById(SecurityUtils.getLoggedUserId());
+        if (user.getRoles().stream().noneMatch(role -> role.getName().equals(UserRole.INDIVIDUAL_USER.name())) && user.getCompanyId() != id) {
+            return ResponseEntity.badRequest().body("No access to resource");
+        }
+        return ResponseEntity.ok().body(userMapper.getById(id));
     }
 
     @Override
@@ -44,24 +46,44 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User saveUser(UserDto userDto) {
-        User user = new User(null, userDto.getName(), userDto.getSurname(), userDto.getEmail(),
-                userDto.getLogin(), passwordEncoder.encode(userDto.getPassword()), null);
-        userMapper.saveUser(user);
-        userMapper.addRoleToUser(userDto.getUserRole().name(), user.getLogin());
-        return getByUsername(user.getLogin());
+    public ResponseEntity saveUser(UserDto userDto) {
+
+        User user = getByUsername(userDto.getLogin());
+        if (user != null) {
+            return ResponseEntity.unprocessableEntity().body("User with login " + user.getLogin() + " already exists");
+        }
+
+        User userToSave = new User(null, userDto.getName(), userDto.getSurname(), userDto.getEmail(),
+                userDto.getLogin(), passwordEncoder.encode(userDto.getPassword()), null, null);
+        userMapper.saveUser(userToSave);
+        userMapper.addRoleToUser(userDto.getUserRole().name(), userToSave.getLogin());
+
+        return ResponseEntity.ok().body(getByUsername(userToSave.getLogin()));
     }
 
     @Override
-    public User updateUser(UserDto userDto, Long userId) {
+    public ResponseEntity updateUser(UserDto userDto, Long userId) {
+
+        if(!userId.equals(SecurityUtils.getLoggedUserId())){
+            return new ResponseEntity<>("No access to resource ", HttpStatus.FORBIDDEN);
+        }
+
         User user = new User(userDto.getName(), userDto.getSurname(), userDto.getEmail());
         userMapper.updateUser(user, userId);
-        return getById(userId);
+
+        return ResponseEntity.ok().body(userMapper.getById(userId));
     }
 
     @Override
-    public void deleteUser(Long userId) {
+    public ResponseEntity deleteUser(Long userId) {
+
+        if(!userId.equals(SecurityUtils.getLoggedUserId())){
+            return new ResponseEntity<>("No access to resource ", HttpStatus.FORBIDDEN);
+        }
+
         userMapper.deleteUserRoles(userId);
         userMapper.deleteUserById(userId);
+
+        return ResponseEntity.ok().build();
     }
 }
